server/requirements.txt
asttokens==2.4.1
bidict==0.22.1
click==8.1.7
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fastapi==0.103.0
h11==0.14.0
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.9
packaging==23.2
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
python-engineio==4.8.2
python-socketio==5.11.0
simple-websocket==1.0.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
uvicorn==0.27.0
wcwidth==0.2.13
wsproto==1.2.0
